# cedar-terminology-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "CEDAR Terminology Server API",
    "title" : "CEDAR Terminology Server API",
    "version" : "2.9.10-SNAPSHOT"
  },
  "paths" : {
...
    "/bioportal/vs-collections" : {
      "get" : {
        "description" : "Find all value set collections.",
        "operationId" : "findAllVSCollections",
        "parameters" : [ {
          "description" : "If true, additional details about each value set collection will be included in the response. Default: false.",
          "in" : "query",
          "name" : "include_details",
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ValueSetCollection"
                  }
                }
              }
            },
            "description" : "Every value set collection, with details only when they were asked for"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Find all value set collections",
        "tags" : [ "Value set collections" ]
      }
    },
    "/bioportal/vs-collections/version-current" : {
      "get" : {
        "description" : "The version triple of a value-set collection's current (\"latest\") locally-stored snapshot — the freeze-on-publish capability for a value-set-valued constraint. Value-set collections are ingested and versioned by the same content-hash mechanism as ontologies. 404 when the collection is not ingested and served locally.",
        "operationId" : "resolveCurrentVersionForValueSetCollection",
        "parameters" : [ {
          "description" : "Value-set collection acronym. Example: CEDARVS.",
          "in" : "query",
          "name" : "collection",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionTriple"
                }
              }
            },
            "description" : "The collection's current version triple"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Value-set collection not resolvable locally"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Resolve the current version triple for a value-set collection",
        "tags" : [ "Value set collections" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CedarError" : {
        "type" : "object",
        "additionalProperties" : true,
        "description" : "A CEDAR error response. Diagnostic fields are populated only when they are relevant to the failure.",
        "properties" : {
          "entities" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named domain entities associated with exception-mapped errors.",
              "nullable" : true
            },
            "description" : "Named domain entities associated with exception-mapped errors.",
            "nullable" : true
          },
          "errorId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "Correlation identifier for an internal error recorded in server logs.",
            "nullable" : true
          },
          "errorKey" : {
            "type" : "string",
            "description" : "Stable machine-readable error identifier.",
            "enum" : [ "templateElementNotCreated", "templateElementNotFound", "templateElementNotDeleted", "templateElementNotUpdated", "templateElementsNotListed", "templateFieldNotCreated", "templateFieldNotFound", "templateFieldNotDeleted", "templateFieldNotUpdated", "templateFieldsNotListed", "templateNotCreated", "templateNotFound", "templateNotDeleted", "templateNotUpdated", "templatesNotListed", "templateInstanceNotCreated", "templateInstanceNotFound", "templateInstanceNotDeleted", "templateInstanceNotUpdated", "templateInstancesNotListed", "noReadAccessToResource", "noWriteAccessToResource", "noReadAccessToFolder", "noWriteAccessToFolder", "noReadAccessToTemplate", "noWriteAccessToTemplate", "noReadAccessToTemplateElement", "noWriteAccessToTemplateElement", "noReadAccessToTemplateField", "noWriteAccessToTemplateField", "noReadAccessToTemplateInstance", "noWriteAccessToTemplateInstance", "noReadAccessToArtifact", "noWriteAccessToArtifact", "noReadAccessToCategory", "noWriteAccessToCategory", "folderNotFound", "artifactNotFound", "artifactPreconditionRequired", "artifactHasMovedOn", "verbatimWriteRefused", "nodeNotFound", "unknownResourceType", "missingParameter", "invalidInput", "sourceFolderNotFound", "sourceResourceNotFound", "targetFolderNotFound", "nodeNotMoved", "groupNotFound", "userNotFound", "uniqueConstraintCollision", "invalidData", "notAuthorized", "permissionMissing", "tokenInvalid", "userInfoLoadByTokenFailed", "userInfoLoadByApiKeyFailed", "cedarUserNotFound", "authorizationNotFound", "permissionNotOwned", "authorizationTypeUnknown", "apiKeyNotFound", "tokenMissing", "tokenExpired", "parentFolderNotSpecified", "parentFolderSpecifiedTwice", "pathNotNormalized", "parentFolderNotFound", "updateInvalidFolderName", "createInvalidFolderName", "nodeAlreadyPresent", "folderNotCreated", "missingNameAndDescription", "missingData", "folderNotDeleted", "artifactNotDeleted", "resourceNotCreated", "invalidResourceType", "invalidArtifactType", "readOtherProfileForbidden", "updateOtherProfileForbidden", "folderCanNotBeDeleted", "folderCanNotBeChanged", "groupAlreadyPresent", "groupCanBeModifiedOnlyByGroupAdmin", "groupCanBeDeletedOnlyByGroupAdmin", "groupMembersCanBeReadOnlyByGroupAdmin", "groupRequiresAdministrator", "groupUsersNotUpdated", "specialGroupCanNotBeDeleted", "folderPermissionsCanNotBeChanged", "unknownInstanceOutputFormat", "folderCopyNotAllowed", "methodNotImplemented", "upstreamServerError", "nothingToDo", "parentCategoryNotFound", "categoryAlreadyPresent", "categoryNotFound", "categoryCanNotBeDeleted", "rootCategoryCanNotBeDeleted", "unableToAttachCategory", "unableToDetachCategory", "noCategoriesWereAttached", "malformedJsonRequestBody", "malformedSearchTerm", "publishedArtifactCanNotBeChanged", "publishedArtifactCanNotBeDeleted", "versioningOnlyOnLatest", "versioningOnlyByOwner", "nonVersionedArtifactType", "createDraftOnlyFromPublished", "publishOnlyDraft", "draftNotCreated", "contentNotValid", "doiNotSupportedByResourceType", "doiCanNotBeSetForEmptyAtId", "doiCanNotBeAltered", "doiCanNotBeSet", "doiAlreadyExists", "dataCiteDOIDisabled", "resourceNotFound", "internalError" ],
            "nullable" : true
          },
          "errorMessage" : {
            "type" : "string",
            "description" : "Alias of message retained for compatibility.",
            "nullable" : true
          },
          "errorReasonKey" : {
            "type" : "string",
            "description" : "More specific machine-readable reason for the error.",
            "enum" : [ "nonEmptyFolder", "nonEmptyCategory", "userHomeFolder", "systemFolder", "templateReferencedInInstances", "validationError", "continuationExpired" ],
            "nullable" : true
          },
          "errorType" : {
            "type" : "string",
            "description" : "Broad category of the error.",
            "enum" : [ "notFound", "invalidArgument", "authorization", "permission", "server", "validationError" ],
            "nullable" : true
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message.",
            "nullable" : true
          },
          "objects" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named structured values associated with the error."
            },
            "description" : "Named structured values associated with the error."
          },
          "operation" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Operation that failed. Its fields depend on the operation type.",
              "nullable" : true
            },
            "description" : "Operation that failed. Its fields depend on the operation type.",
            "nullable" : true
          },
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named scalar values associated with the error."
            },
            "description" : "Named scalar values associated with the error."
          },
          "status" : {
            "type" : "string",
            "description" : "Symbolic HTTP response status.",
            "enum" : [ "BAD_REQUEST", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "CONFLICT", "PRECONDITION_FAILED", "UNSUPPORTED_MEDIA_TYPE", "UNPROCESSABLE_ENTITY", "PRECONDITION_REQUIRED", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "HTTP_VERSION_NOT_SUPPORTED" ]
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Numeric HTTP response status.",
            "maximum" : 599,
            "minimum" : 400
          },
          "suggestedAction" : {
            "type" : "string",
            "description" : "Recovery action suggested to the client.",
            "enum" : [ "none", "requestRole", "logout", "logoutImmediately", "provideAuthorizationHeader", "refreshToken" ],
            "nullable" : true
          }
        },
        "required" : [ "status", "statusCode" ]
      },
...
      "ValueSetCollection" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string"
          },
          "@type" : {
            "type" : "string"
          },
          "details" : {
            "$ref" : "#/components/schemas/ValueSetCollectionDetails"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        }
      },
      "ValueSetCollectionDetails" : {
        "type" : "object",
        "properties" : {
          "categories" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "creationDate" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "documentation" : {
            "type" : "string"
          },
          "hasOntologyLanguage" : {
            "type" : "string"
          },
          "homepage" : {
            "type" : "string"
          },
          "numberOfClasses" : {
            "type" : "integer",
            "format" : "int32"
          },
          "publication" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "released" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          }
        }
      },
...
      "VersionTriple" : {
        "type" : "object",
        "properties" : {
          "declaredVersion" : {
            "type" : "string"
          },
          "effectiveDate" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}